Add support for multiple images
This patch adds support for running tests on multiple images. Instead of
parsing just 1 image, it attempts to always parse multiple images from
config files and/or CLI.

To run dox on multiple images from the CLI, you'll have to:

    $ dox -i infra/trusty,infra/f20

The tox.ini `docker:images` option is now a comma-separated list of
images, whereas the dox.yml one is an array of images.

Change-Id: I124c299b208be5c31c63187af754705c48de5693