Add option to background test container
It may be handy to have dox build and start a container that
would run some sort of server process, such as a web server
or other long-lived process. In this case, we would need to
daemonize the container with the docker -d option. This change
adds a -b/--background option to dox to do that.

Change-Id: I755fcb6bb98a20ca7511d69b44fb7fe50d34dcf5