Add bindep.txt support
Bindep is a tool for checking the presence of binary packages needed
to use an application / library. It started life as a way to make it
easier to set up a development environment for OpenStack projects.
While OpenStack depends heavily on pip for installation of Python
dependencies, some dependencies are not Python based, and particularly
for testing, some dependencies have to be installed before pip can be
used - such as virtualenv and pip itself.

Change-Id: I92cbfc36610cb931790b6e607dcb317ce2dac08e
Signed-off-by: Paul Belanger <pabelanger@redhat.com>